Split browse routes to own file
