+++ CHANGELOG-reorg-routes.md
@@ -1,0 +1,1 @@
+- In the python code, split the `/browse/*` routes to their own file, to keep things reasonable.
