+++ context/app/main.py
@@ -1,6 +1,6 @@
 from flask import Flask, session, render_template
 
-from . import routes_main, routes_auth, routes_cells, routes_markdown, default_config
+from . import routes_main, routes_browse, routes_auth, routes_cells, routes_markdown, default_config
 from .flask_static_digest import FlaskStaticDigest
 flask_static_digest = FlaskStaticDigest()
 
@@ -52,6 +52,7 @@ def create_app(testing=False):
     flask_static_digest.init_app(app)
 
     app.register_blueprint(routes_main.blueprint)
+    app.register_blueprint(routes_browse.blueprint)
     app.register_blueprint(routes_cells.blueprint)
     app.register_blueprint(routes_auth.blueprint)
     app.register_blueprint(routes_markdown.blueprint)
