+++ context/app/routes_markdown.py
@@ -4,12 +4,12 @@ from os.path import dirname
 
 from flask import Blueprint, render_template, request, redirect
 
-from .routes_main import get_default_flask_data
+from .utils import get_default_flask_data
 
 # NOTE: A better approach might be to look again at the handful of libraries
 # that handle this, or to pre-render everything when flask starts.
 
-blueprint = Blueprint('routes_markdown', __name__, template_folder='templates')
+blueprint = Blueprint('markdown', __name__, template_folder='templates')
 
 
 def _title_from_md(md):
