+++ context/app/utils.py
@@ -1,0 +1,15 @@
+from flask import current_app
+
+
+def get_default_flask_data():
+    return {
+        'endpoints': {
+            'elasticsearchEndpoint': current_app.config['ELASTICSEARCH_ENDPOINT']
+            + current_app.config['PORTAL_INDEX_PATH'],
+            'assetsEndpoint': current_app.config['ASSETS_ENDPOINT'],
+            'entityEndpoint': current_app.config['ENTITY_API_BASE'],
+            'xmodalityEndpoint': current_app.config['XMODALITY_ENDPOINT'],
+            'gatewayEndpoint': current_app.config['GATEWAY_ENDPOINT'],
+        },
+        'globalAlertMd': current_app.config.get('GLOBAL_ALERT_MD')
+    }
